Refactor files structers
