+++ src/pages/Domains/Create.tsx
@@ -1,10 +1,10 @@
 import { Button, Drawer, FormProps } from "antd";
-import useCreateNotification from "../hooks/useCreateNotification";
-import DomainForm from "./DomainForm";
-import { useCreateDomainMutation } from "../state/domains/domainsApiSlice";
 import { useEffect, useState } from "react";
+import { useCreateDomainMutation } from "../../state/domains/domainsApiSlice";
+import useCreateNotification from "../../hooks/useCreateNotification";
+import DomainForm from "../../components/forms/DomainForm";
 
-export default function CreateDomain({ callBack }: { callBack: () => void }) {
+export default function Create({ callBack }: { callBack: () => void }) {
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
   const { notify, contextHolder } = useCreateNotification();
   const [createDomain, { isSuccess, isLoading, isError }] =
